refactor Batching strategy
