+++ Flagship/Api/BatchingContinuousCachingStrategy.cs
@@ -35,6 +35,11 @@ namespace Flagship.Api
             }
             Logger.Log.LogDebug(Config, string.Format(HIT_ADDED_IN_QUEUE, JsonConvert.SerializeObject(hit.ToApiKeys())), ADD_HIT);
 
+            if (HitsPoolQueue.Count>= Config.TrackingMangerConfig.PoolMaxSize)
+            {
+                _ = SendBatch(CacheTriggeredBy.BatchLength);
+            }
+
         }
 
         public override async Task NotConsent(string visitorId)
@@ -53,117 +58,8 @@ namespace Flagship.Api
             await FlushHitsAsync(keys);
         }
 
-        public override async Task SendBatch()
-        {
-            if (ActivatePoolQueue.Any())
-            {
-                var activateHits = ActivatePoolQueue.Values.ToList();
-                var keys = activateHits.Select(x => x.Key);
-                foreach (var item in keys)
-                {
-                    ActivatePoolQueue.Remove(item);
-                }
-                await SendActivate(activateHits, null);
-            }
-            var batch = new Batch()
-            {
-                Config = Config
-            };
-
-            var count = 0;
-
-
-            var hitKeysToRemove = new List<string>();
-
-            foreach (var item in HitsPoolQueue)
-            {
-                count++;
-                var batchSize = JsonConvert.SerializeObject(batch).Length;
-                if (batchSize > Constants.BATCH_MAX_SIZE || count > Config.TrackingMangerConfig.PoolMaxSize)
-                {
-                    break;
-                }
-                batch.Hits.Add(item.Value);
-                hitKeysToRemove.Add(item.Key);
-            }
-
-            if (!batch.Hits.Any())
-            {
-                return;
-            }
-
-            foreach (var key in hitKeysToRemove)
-            {
-                HitsPoolQueue.Remove(key);
-            }
-
-            var requestBody = batch.ToApiKeys();
-
-            try
-            {
-                var requestMessage = new HttpRequestMessage(HttpMethod.Post, Constants.HIT_EVENT_URL);
-
-                requestMessage.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue(Constants.HEADER_APPLICATION_JSON));
-
-                var postDatajson = JsonConvert.SerializeObject(requestBody);
-
-                var stringContent = new StringContent(postDatajson, Encoding.UTF8, Constants.HEADER_APPLICATION_JSON);
-
-                requestMessage.Content = stringContent;
-
-                var response = await HttpClient.SendAsync(requestMessage);
-
-                if (response.StatusCode != System.Net.HttpStatusCode.OK)
-                {
-                    var message = new Dictionary<string, object>()
-                    {
-                        {"StatusCode:", response.StatusCode},
-                        {"ReasonPhrase", response.ReasonPhrase },
-                        {"response", await response.Content.ReadAsStringAsync() }
-                    };
-
-                    throw new Exception(JsonConvert.SerializeObject(message));
-                }
-
-                Logger.Log.LogDebug(Config, string.Format(BATCH_SENT_SUCCESS, JsonConvert.SerializeObject(requestBody)), SEND_BATCH);
-
-                await FlushHitsAsync(hitKeysToRemove.ToArray());
-            }
-            catch (Exception ex)
-            {
-                foreach (var item in batch.Hits)
-                {
-                    HitsPoolQueue[item.Key] = item;
-                }
-                Logger.Log.LogError(Config, Utils.Utils.ErrorFormat(ex.Message, new
-                {
-                    url = Constants.HIT_EVENT_URL,
-                    body = requestBody
-                }), SEND_BATCH);
-            }
-        }
-
-        public async override Task ActivateFlag(Activate hit)
-        {
-
-            var hitKey = $"{hit.VisitorId}:{Guid.NewGuid()}";
-            hit.Key = hitKey;
-            var activateHitPool = new List<Activate>();
-            if (ActivatePoolQueue.Any())
-            {
-                activateHitPool = ActivatePoolQueue.Values.ToList();
-                var keys = activateHitPool.Select(x => x.Key);
-                foreach (var item in keys)
-                {
-                    ActivatePoolQueue.Remove(item);
-                }
-            }
-
-            await SendActivate(activateHitPool, hit);
-
-        }
-
-        protected async override Task SendActivate(ICollection<Activate> activateHitsPool, Activate currentActivate)
+       
+        protected async override Task SendActivate(ICollection<Activate> activateHitsPool, Activate currentActivate, CacheTriggeredBy batchTriggeredBy)
         {
             var url = Constants.BASE_API_URL + URL_ACTIVATE;
             var requestMessage = new HttpRequestMessage(HttpMethod.Post, url);
@@ -181,6 +77,7 @@ namespace Flagship.Api
             }
 
             var requestBody = activateBatch.ToApiKeys();
+            var now = DateTime.Now;
 
             try
             {
@@ -210,7 +107,9 @@ namespace Flagship.Api
                     await FlushHitsAsync(hitKeysToRemove);
                 }
 
-                Logger.Log.LogDebug(Config, string.Format(HIT_SENT_SUCCESS, postDatajson), SEND_ACTIVATE);
+                requestBody["duration"] = (DateTime.Now - now).TotalMilliseconds;
+                requestBody["batchTriggeredBy"] = $"{batchTriggeredBy}";
+                Logger.Log.LogDebug(Config, string.Format(HIT_SENT_SUCCESS, JsonConvert.SerializeObject(requestBody)), SEND_ACTIVATE);
             }
             catch (Exception ex)
             {
@@ -233,7 +132,9 @@ namespace Flagship.Api
                             {Constants.HEADER_X_SDK_CLIENT, Constants.SDK_LANGUAGE },
                             {Constants.HEADER_X_SDK_VERSION, Constants.SDK_VERSION }
                         },
-                    body = requestBody
+                    body = requestBody,
+                    duration = (DateTime.Now - now).TotalMilliseconds,
+                    batchTriggeredBy = $"{batchTriggeredBy}"
                 }), SEND_ACTIVATE);
             }
         }
